My fighter header (#336)
* added almost full fighter.h Fighter struct, including some info from decomping

* created fighter.c with all fighter.s funcs as inline asm

* added some fighter functions

* comment

* added decomp link for func_80068354

* reduced fighter.h to used data only

* synched Fighter struct with master

* added fillers

* spaces -> tabs

* removed jobj x123 member offsets

* linked to snuffysasa's fix for func_80068354

* added func_80068354

* cleanup

* cleanup

* procUpdate integration

* now fighter_procUpdate compiles

* more cleanup. procUpdata doesn't OK with the new fighter.h for some reason

* updated decomp link

* tabs to spaces to get better diffs to master

* attr usage comment

* tabs > spaces

* removed unused HitVictim struct

* reformatted Fighter::attr members

* changed func_800693AC param types

* linked snuffysasa's match of func_8006C0F0

* fixed func_80067C98 using addi instead of mr

* move fpclassify from fighter.c to math.h

* moved fighter.c extern funcs to functions.h

* created header only fighter branch

* replaced FighterID with more accurate FighterKind and CharacterKind enums from akaneia

* re-add fighter.c, and updates to fighter.h, and lots of function fixes, and add inlines to jobj.h

* change line numbers from hex to decimal

* small fix to player.c

* include player.h for functions

* update some of kaluas functions, use the jobj inlines, move all headers to fighter.h for easier context generation

* add func_800686E4

* add func_8006876C, and update fighterBone to have jobj pointers

* add new functions in C, update the fighter struct to have substructs phys and dmg and needed to update functions in many files for new structure

* change player function output

* another fix to player interface

* fix stageinfo type

* matching lots of functions in fighter.c, lots of updates to fighter structs in fighter.h as merging different matches, and updating old matches to the latest fighter context

* make setTranslate one function again to match master, and is correct, fixed two functions

* move fighter code back to asm

* more small tweaks to fighter structs

* small fix to func_800868D4

* one more tweak to fighter struct and now the dol is matching

* remove commented out fighter.c

* remove the inner phys struct and rename a var, based on rtburns comment

* more updates to fighter header

* more tweaks to x2229 field

* add offsets to struct fields in fighter and little cleanup

* update fighter.h with updated function declaraions

* match several more smaller functions in fighter.c

+++ src/melee/pl/player.c
@@ -1449,7 +1449,7 @@ void Player_UnsetFlagsBit1(s32 slot) {
     player->flags.b1 = zero;
 }
 
-u8 Player_GetFlagsBit3(s32 slot) {  
+s32 Player_GetFlagsBit3(s32 slot) {  
     StaticPlayer* player;
     u8 bit3;
     Player_CheckSlot(slot);
